# packages/contracts/openapi/merchant.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProxyShop merchant service",
    "version": "1.0.0",
    "description": "Code minting, Shopify webhooks, the pixel collector, and envelope administration. The envelope routes are SEALED state (C3/S7): the exchange has no grant to read them and `.importlinter` forbids the import."
  },
...
  "paths": {
...
    "/stores/{store_id}/bids/solicit": {
      "post": {
        "operationId": "solicitStoreAgentBid",
        "summary": "Ask this store's own agent for a bid, and record what it answered.",
        "description": "This is how R9's kill switch becomes OBSERVABLE. The merchant service posts a real `BidRequest` to the store agent's published `POST /v1/bid-requests` - the same door the exchange solicits through - and records the answer verbatim: a `Bid`, or the contract's 204 decline whose reason travels in `x-proxyshop-decline-reason` (`store_killed`, `envelope_not_activated`, ...). Kill the store, solicit again, and the row that comes back is the proof. It is a REHEARSAL and every row says so (`origin: dashboard`); it is not the live auction log, whose per-auction rationale lives in the agent's own in-process shadow log and has no served route. The body is optional: anything omitted is defaulted from the store's own envelope, the cluster from `pursue_clusters`, so a rehearsal is not answered `cluster_not_pursued` for a cluster the merchant never asked to pursue. The pseudonym is a fixed synthetic value and never a buyer's (R5).",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cluster_id": {
                    "type": "string"
                  },
                  "query": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "ship_to": {
                    "type": "string"
                  },
                  "currency": {
                    "type": "string"
                  },
                  "budget_band": {
                    "type": "string"
                  },
                  "hard_constraints": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "preferences": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                },
                "additionalProperties": false
              },
              "example": {
                "cluster_id": "cluster-warm-layers",
                "query": "a warm mid-layer for cold commutes",
                "hard_constraints": [
                  {
                    "field": "material",
                    "op": "eq",
                    "value": "merino wool"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The agent answered. A decline is a 200: the solicitation succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "recorded_at",
                    "origin",
                    "store_id",
                    "auction_id",
                    "outcome",
                    "activation",
                    "may_bid",
                    "claims",
                    "contradiction"
                  ],
                  "properties": {
                    "recorded_at": {
                      "type": "string"
                    },
                    "origin": {
                      "type": "string",
                      "enum": [
                        "dashboard"
                      ]
                    },
                    "store_id": {
                      "type": "string"
                    },
                    "auction_id": {
                      "type": "string"
                    },
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "bid",
                        "declined"
                      ]
                    },
                    "activation": {
                      "$ref": "https://proxyshop.dev/schemas/protocol.schema.json#/$defs/EnvelopeActivation"
                    },
                    "may_bid": {
                      "type": "boolean"
                    },
                    "claims": {
                      "type": "integer"
                    },
                    "contradiction": {
                      "type": "boolean",
                      "description": "The merchant stopped this store and its agent bid anyway. Reachable today: the merchant service holds the authoritative envelope and nothing delivers it to a separately deployed agent, which reads its own STORE_AGENT_CONTEXT once at start-up."
                    },
                    "decline_reason": {
                      "type": "string"
                    },
                    "offer": {
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "recorded_at": "2026-01-01T00:00:00+00:00",
                  "origin": "dashboard",
                  "store_id": "store-alpha",
                  "auction_id": "dashboard-rehearsal-0000",
                  "outcome": "declined",
                  "activation": "killed",
                  "may_bid": false,
                  "claims": 0,
                  "contradiction": false,
                  "decline_reason": "store_killed",
                  "detail": "the agent answered the contract's decline; the reason is in its header."
                }
              }
            }
          },
          "400": {
            "description": "The body was present and is not a JSON object.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unparseable-body",
                        "not-a-solicitation"
                      ]
                    },
                    "detail": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "unparseable-body"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the wrong one.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized"
                      ]
                    }
                  }
                },
                "example": {
                  "error": "unauthorized"
                }
              }
            }
          },
          "413": {
            "description": "The body exceeded this service's cap.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "body-too-large"
                      ]
                    }
                  }
                },
                "example": {
                  "error": "body-too-large"
                }
              }
            }
          },
          "502": {
            "description": "The store agent's address is configured and nothing answered it.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "store-agent-unreachable"
                      ]
                    },
                    "detail": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "store-agent-unreachable",
                  "detail": "the store agent at store-agent:8085 did not answer: All connection attempts failed. The address is configured; the service is not up."
                }
              }
            }
          },
          "503": {
            "description": "No admin token, or no address for the store's own agent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "admin-api-not-configured",
                        "store-agent-not-configured"
                      ]
                    },
                    "missing": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "detail": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "store-agent-not-configured",
                  "missing": [
                    "STORE_AGENT_URL"
                  ],
                  "detail": "this deployment states no address for the store's own agent"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "The merchant admin token, configured as `MERCHANT_ADMIN_TOKEN`. The envelope routes are SEALED state (C3/S7) and refuse every caller — with 503, not 200 — until it is set."
      }
    }
  }
}